move terraform to own pkg to keep hcl out of lib/services import graph
