+++ lib/accesslists/terraform/testdata/TestGenerateTerraformConfigWithPresetBuilder/access-list-nil-roles.golden
