+++ lib/accesslists/terraform/testdata/TestGenerateTerraformConfigWithPresetBuilder/access-roles-nil-access-list.golden
