+++ lib/accesslists/terraform/testdata/TestGenerateTerraformConfigWithPresetBuilder/long-term-preset.golden
