+++ lib/accesslists/terraform/testdata/TestGenerateTerraformConfigWithPresetBuilder/short-term-preset.golden
